chore(landing): reorganize component files and update imports for consistency
