+++ apps/demo/tsconfig.app.json
@@ -21,5 +21,5 @@
     "vite.config.ts",
     "vitest.config.ts"
   ],
-  "include": ["src/**/*.js", "src/**/*.jsx", "src/**/*.ts", "src/**/*.tsx", "../../libs/shell/src/lib/nav/NavMain.tsx", "../../libs/shell/src/lib/nav/NavUser.tsx", "../../libs/shell/src/lib/nav/TeamSwitcher.tsx", "../../libs/landing/src/lib/components/AboutSection.tsx", "../../libs/landing/src/lib/components/CommunitySection.tsx", "../../libs/landing/src/lib/components/DemoSection.tsx", "../../libs/landing/src/lib/components/features-section/FeaturesSection.tsx", "../../libs/landing/src/lib/components/Footer.tsx", "../../libs/landing/src/lib/components/GetStartedSection.tsx", "../../libs/landing/src/lib/components/HeroSection.tsx"]
+  "include": ["src/**/*.js", "src/**/*.jsx", "src/**/*.ts", "src/**/*.tsx", "../../libs/shell/src/lib/nav/NavMain.tsx", "../../libs/shell/src/lib/nav/NavUser.tsx", "../../libs/shell/src/lib/nav/TeamSwitcher.tsx", "../../libs/landing/src/lib/components/about-section/AboutSection.tsx", "../../libs/landing/src/lib/components/community-section/CommunitySection.tsx", "../../libs/landing/src/lib/components/demo-section/DemoSection.tsx", "../../libs/landing/src/lib/components/features-section/FeaturesSection.tsx", "../../libs/landing/src/lib/components/footer-section/Footer.tsx", "../../libs/landing/src/lib/components/get-started-section/GetStartedSection.tsx", "../../libs/landing/src/lib/components/hero-section/HeroSection.tsx"]
 }
