+++ libs/landing/src/lib/components/about-section/AboutSection.tsx
