+++ libs/landing/src/lib/components/community-section/CommunitySection.tsx
