+++ libs/landing/src/lib/components/footer-section/Footer.tsx
