+++ libs/landing/src/lib/components/get-started-section/GetStartedSection.tsx
