+++ libs/landing/src/lib/components/hero-section/HeroSection.tsx
