+++ libs/landing/src/lib/components/index.ts
@@ -1,8 +1,8 @@
-export * from './AboutSection';
-export * from './CommunitySection';
-export * from './CTASection';
-export * from './DemoSection';
+export * from './about-section/AboutSection';
+export * from './community-section/CommunitySection';
+export * from './cta-section/CTASection';
+export * from './demo-section/DemoSection';
 export * from './features-section';
-export * from './HeroSection';
-export * from './Footer';
-export * from './GetStartedSection';
+export * from './hero-section/HeroSection';
+export * from './footer-section/Footer';
+export * from './get-started-section/GetStartedSection';
